CU-86947ja9y: Move the fix to a separate utils module along with the default weighted average function (to avoid circular imports)
